test(spanner): move testing helper to a test
